A few more core changes to accept our front-end.
Rewrote LIBS: in the makefile to handle SDL Wii properly.

The emulator now compiles and outputs executables which attempt to load a ROM from "sd:/boot.nds" and then emulate it. We've tested a handful of ROMs but none of them work, so we are now working through a long debugging process. 

git-svn-id: http://desmumewii.googlecode.com/svn/trunk@13 dc052594-c8ac-11de-afa3-455a29e9879b