Added code for a special console window from dhewg.
We use the console window to finally display text. However, we have to move SLOWLY or else we will cause a crash and never get any useful info beyond a black screen. Thus, I've placed an infinite loop to demonstrate that output is fixed (but again, many bugs not perfected)

Please see the source code for more details.

git-svn-id: http://desmumewii.googlecode.com/svn/trunk@21 dc052594-c8ac-11de-afa3-455a29e9879b