The console is back baby! It seems that it didn't like having the SDL_video being set for the DS first. This will have to be changed, but for now, at least we can debug.
It also saddens me to no end to point out that the problem that plagued the first DesmumeWii is back in full force. Yes, the bane of my existence, MMU.cpp is where everything stops. I still do not know what the problem is, or how to fix it.

I put code into NDSSystem to make it return to the main menu before calling the MMU code. If you see "NDS_Init done" as the last console trace, then everyone is on the same page. Comment it out to try to fix MMU. May the gaming gods have mercy on our souls.

git-svn-id: http://desmumewii.googlecode.com/svn/trunk@47 dc052594-c8ac-11de-afa3-455a29e9879b