Heven't tested it but we should now have input working.
Controls are as follows:
The classic controller works as though it is the DS. All the buttons on the classic controller match up to their DS counter parts (A = A, B = B, etc.)

Holding A on the Wiimote signals that the stylus is touching the screen. While holding A, press the D-Pad on the Wiimote to move the position of the stylus.

Better controls in the future hopefully.

Also hoping someone can test the space invaders game ( remember you MUST have the classic controller) and tell me if these controls work.

git-svn-id: http://desmumewii.googlecode.com/svn/trunk@54 dc052594-c8ac-11de-afa3-455a29e9879b