Create default firmware.
Super Princess Peach (E) was failing leading to a white screen because copy_firmware_user_data() in NDSsystem.cpp was failing. So this is setting default firmware.

This does not fix the problem, the loading just gets further. I'm tired and going to bed so if someone else wants to verify it while I sleep then please do.

git-svn-id: http://desmumewii.googlecode.com/svn/trunk@106 dc052594-c8ac-11de-afa3-455a29e9879b