I've come to the realization that the r85 filebrowser is fundamentally flawed. Lengthy and detailed testing by several people has revealed that it is the source of the white screen issue and the source of a code dumps which occur after selecting SD or USB.
What I've done here is 50% fix the first of many flaws. Now, if directory parameter of the filebrowser is not located in the device, the filebrowser will jump to the root of the external device. However, any ROMs selected this way will not run. Your guess is as good as mine why :/ I would finish what I started but I've been working on it for several hours and can't put any more time into it today.

git-svn-id: http://desmumewii.googlecode.com/svn/trunk@113 dc052594-c8ac-11de-afa3-455a29e9879b