We have GX!
Back in March I made a joke about learning GX. Well guess what? I managed to port (or rather, redo) the OGLRender code into native GX.

Now, the textures are ALL wrong, and yes there are four of every sprite. I address this in the Issues_with_GX page on the wiki. This is still HIGHLY EXPERIMENTAL so do not expect perfection right out of the gate. 

You can choose to use GX on the same screen that you choose USB/SD card. Up and down control this.

git-svn-id: http://desmumewii.googlecode.com/svn/trunk@187 dc052594-c8ac-11de-afa3-455a29e9879b